Move away from uuid tokens
uuid tokens aren't guaranteed to be cryptographically secure, therefore
moving to randomly generated bytes is preferable. This is a breaking
change that involved rewriting historical migration files, and so the version
number has been bumped to 0.2.0.

For future breaking changes I'll write proper migrations but since as
far as I'm aware nobody is using this project yet I decided to just
"rewrite history" as it were.